Navbar component and general cleanup
