+++ client/src/components/Navbar/style.scss
@@ -1,0 +1,21 @@
+nav{
+    display:flex;
+    margin-top:10px;
+    padding-bottom:10px;
+    border-bottom:solid grey 2px;
+    width:100%;
+}
+
+div.level-item{
+    display:flex;
+}
+
+#nav-logo{
+    height:45px;
+}
+
+// a .navbar-burger{
+//     width:100%;
+// }
+
+// img{width:100000px}
